refactor: move SimRandomizeTester
